+++ src/main/kotlin/com/github/gllxl/imagepreview/CssLineMarkerContributor.kt
@@ -1,9 +1,11 @@
 package com.github.gllxl.imagepreview
 
 import com.intellij.execution.lineMarker.RunLineMarkerContributor
+import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.css.impl.CssElementTypes.*
 import com.intellij.psi.util.elementType
+import com.intellij.refactoring.suggested.startOffset
 
 fun getCSSBackgroundVariableContent (element: PsiElement): String? {
   return element.children.find { it.elementType === CSS_STRING }?.text
@@ -15,6 +17,11 @@ class CssLineMarkerContributor : LineMakerContributor() {
       val imageUrl = getCSSBackgroundVariableContent(element)
 
       if (imageUrl != null ) {
+
+        val document = PsiDocumentManager.getInstance(element.project).getDocument(element.containingFile) ?: return null
+        val lineNumber = document.getLineNumber(element.startOffset)
+        ImageMapping.setLineMapping(element.containingFile.virtualFile, lineNumber, removeUrlQuotes(imageUrl))
+
         return getLineMaker(removeUrlQuotes(imageUrl))
       }
       return null