+++ src/main/kotlin/com/github/gllxl/imagepreview/ImageMapping.kt
@@ -1,0 +1,33 @@
+package com.github.gllxl.imagepreview
+
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.util.net.HttpConfigurable
+import java.awt.image.BufferedImage
+import java.net.HttpURLConnection
+import java.util.HashMap
+import javax.imageio.ImageIO
+
+object ImageMapping {
+
+  private val pool = HashMap<VirtualFile, HashMap<Number, String>>()
+
+  private fun hasFileInPools(file: VirtualFile): Boolean {
+    return pool.containsKey(file)
+  }
+
+  fun setLineMapping(file: VirtualFile, lineNumber: Number, imgUrl: String) {
+    val inHashMap = HashMap<Number, String>()
+    inHashMap[lineNumber] = imgUrl
+    val target = pool[file]
+    if (hasFileInPools(file)){
+      target?.put(lineNumber,imgUrl)
+    } else {
+      pool[file] = inHashMap
+    }
+  }
+
+  fun getLineMapping(file: VirtualFile, lineNumber: Number): String? {
+    return pool[file]?.get(lineNumber)
+  }
+
+}